Sonar fixes

+++ src/items/bed.cpp
@@ -210,8 +210,8 @@ void BedItem::regeneratePlayer(Player* player) const
 	if (condition != nullptr) {
 		uint32_t regen;
 		if (condition->getTicks() != -1) {
-			regen = convertToSafeInteger<uint32_t>(std::min<int32_t>((condition->getTicks() / 1000), sleptTime) / 30); // RATE_HEALTH_REGEN_SPEED and RATE_MANA_REGEN_SPEED?
-			const int32_t newRegenTicks = condition->getTicks() - (regen * 30000);
+			regen = convertToSafeInteger<uint32_t>(std::min<int64_t>((condition->getTicks() / (int64_t)1000), (int64_t)sleptTime) / 30); // RATE_HEALTH_REGEN_SPEED and RATE_MANA_REGEN_SPEED?
+			const int64_t newRegenTicks = condition->getTicks() - (regen * 30000);
 			if (newRegenTicks <= 0) {
 				player->removeCondition(condition);
 			} else {
